cli/daemon/run: ignore changes to non-Go files
We previously live-reloaded optimistically, in case there was a go:embed
of a non-Go file or something like it. This turned out to cause a lot of
spurious reloads, particularly when working with JavaScript projects
in the same repository.

For now only trigger live-reload when Go-related files change.
We can make this more sophisticated in the future, if desired.