cli/daemon/run: fix panic when racing with proc start
We add the run to the manager's map before it has fully started
in order to avoid spurious "run not found" errors.

This can in rare circumstances create another race,
where we try to access the proc before it has started.
We already guard against this in most places but forgot one.