compiler: parser: cli: support custom API paths and methods
This adds support for custom API paths and HTTP methods for non-raw endpoints.

When you define a non-raw endpoint, the path parameters must correspond
in position and name to additional function parameters. Like so:

	//encore:api public path=/orders/:id
	func GetOrder(ctx context.Context, id int) (*Order, error)

The type of the function parameter is used for validation.

Calling for example /orders/foo would result in a 400 Bad Request
since "foo" is not an integer.

The function parameter types can be strings, integers, booleans,
or UUIDs (using encore.dev/types/uuid.UUID).