Simplify contribution instructions
This change simplifies the instructions for contributing to Encore
and building from source.

It introduces a new env variable, `ENCORE_DAEMON_DEV=1` that causes
the Encore daemon to use a fixed port for the daemon dashboard (`9400`)
and updates the development dashboard frontend to use it when in
development mode.

It also introduces a new `encore daemon env` command that prints Encore
environment configuration for making it easier to find the location
of the pre-built Encore Go runtime.