[ENC-310] Update DashApp to Support Generics (#5)
* [ENC-310] Add Support for Generics Client Side

This commit converts `est.Param` to contain a `*schema.Type` rather
than a `*schema.Decl`, as we now need to know what type arguments have
been given to initialise a type, where as a `Decl` is only the raw
underlying delecation of that type.

I've then updated the Go based go generator for a typescript client
to have full support for generics. While I was there I refactored
how the test is run, which makes it easier to both edit the input/output
of the test, but also add more languages in the future by simply adding an
`expected_langname.ext` file to the `testdata` folder.

This commit starts to add support for generics on the client side by
updating the proto code generator to also generate typescript files,
which we can then include directly from the dashapp. This minimises
the risk of the two systems (Go + Typescript) becoming different in
terms of their representation of the Encore app.

* [ENC-310] Update the dashapp to support generics

This commit updates the `dashapp` so it works with generic types.
While I was here I removed some duplication of code in the JSON
default parameters for requests. As part of unifiying the dupliated
code I've moved the SchemaView over to using `CodeMirror` as well
so we get syntax highlighting in the API docs.

I've also updated the response renderers in both `RPCCaller` and
the trace overview to also render the payloads with syntax highlighting.

* Addressing PR comments