cli/daemon/sqldb: adjust sql cluster setup to mimic production environment more closely
Locally, there used to be only the 'encore' user that had superuser
privileges. The change mimics the production setup, where there's a
'postgres' user with superuser privileges and an encore user as table
owner with "Create Database" privileges.

This change allows implementing RowLevelSecurity(RLS) within encore and
have it work locally the same way as it does in the cloud.