cli: compiler: runtime: merge cloud runtime and local runtime
This commit combines the two runtime implementations Encore has provided
into a single runtime. This is done in order to simplify future
development by reducing the amount of duplication necessary to implement
new functionality.

APIs internal to Encore are hidden from the API docs through the use
of the "encore_internal" build tag, which the compiler now always
provides. Additionally, in order to switch on/off certain functionality
for local development only there is now a "encore_local" build tag
which is passed in by default by the daemon when compiling for
local development.

Finally to bring the runtimes to parity with each other we now
generate auth keys in the daemon when starting up the app locally,
in order to exercise the production code paths. Similarly the local
runtime outputs JSON logs instead of preformatted console logs,
and the conversion to console formatting happens in the CLI.