Add App & Request Metadata
In this PR we've added the code for returning the CurrentOp() and Meta() API's as defined in #171. Which adds the ability for running Encore applications to retrieve metadata about the application itself and the current request.

This commit also updates the healthz endpoint to include information about the deployment, which can then be used by Encore's infrastructure system to verify the deployment of new binaries has been successful on all platforms (i.e the old binary isn't still being served).

I've also merged the two runtime config files together, by getting the compiler
to use the combined runtime. This allows us to ensure that the two config's structs
can never get out of sync; as now there is just one!