runtime/beta/errs: remove encore_internal build tag
It's causing annoying build errors when using this module for
use cases outside of building an Encore app, such as referencing
the config structs.

Having this build tag is not important for now as the public API is
still in a separate module. If we go ahead with merging that into this
repo we can revisit this decision.