runtime: separate sql server config from databases
It's common to use the same database server with multiple databases.
In combination with client side TLS certificates this creates an
enormous amount of duplicated data, to the point where the size of the
encoded config can exceed the 64KiB limits imposed by Cloud Run.

Fix this by separating out the SQL Server configuration from the
SQL Database configuration, so that multiple databases on the same
server can use the same TLS configuration.