cli/cmd/go-encore-wrapper: new command
This adds a new command, go-encore-wrapper, that is designed
to wrap the "go" command and forward certain commands to
the "encore" CLI, like running tests.

The purpose is to build a better integration with tools like
vscode-go that only support replacing the "go" tool with
another tool that has the same command structure.

With this change, users will be able to configure vscode-go
to use Encore when appropriate simply by defining:

    "go.alternateTools": {"go": "go-encore-wrapper"}