Add `encoding.DescribeRPC`
This commit adds a new `DescribeRPC` function to the encoding package.
This function's job is given an RPC to describe how to encode requests
and responses for that RPC on the wire.

The output of this function can then be used for both the client generation
and for the code generation inside Encore request handlers for applications.