Add encoding.DescribeResponse
This commit adds a new `DescribeResponse` function to the encoding package.
This function's job is to describe the http wire representation of a Response type

The output of this function can then be used for both the client generation
and for the code generation inside Encore request handlers for applications.