Shared Go Code Generator for Marshalling
This commit ports all the `typeDecoder` logic from compiler/internal/codegen/codegen_main.go
into a new shared package called `gocodegen` and adds a new set of methods for encoding to
strings for use in the Go client.

I've not removed or modified the existing compiler to use this new package yet. But that
will be a future commit once we've finished adding the header support to the compiler.