Add Header support to Go Client Generator
This commit adds header support to the Go Client Generator using the new
DescribeRPC and shared Type Marshaling generator.

The golden package is now being used for the Client Generator to make it easier
to update the expected test files.

It updates the input to the client generator test to include a bunch of new
cases to try and expose as many test edges as possible.