compiler/internal/codegen: Add native encore support for headers (#222)
This change introduces the `header` field tag, which can be used in encore request and response objects to encode the field as a http header. We've also added support for query string parameters in POST requests. Encore will try to read a request field from the query string if it is tagged with either the `qs` or `query` tag.

Example:

```
type Params struct {
	header string `header:"x-header"`
	query   string `query:"query"`
	body    string `body:"body"`
}

//encore:api public method=POST
func Echo(ctx context.Context, params *Params) (*Params, error) {
	return params, nil
}
```

The example above will:
* read `header` from the `x-header` field of the request HTTP header
* read `query` from the request query string
* read `body` as a json field from the request body
* write `header` as `x-header` in the response HTTP header
* write `query` as a json field in the response body
* write `body` as a json field in the respinse body