Add end-to-end tests for Typescript Client
This commit adds end to end tests for the generated Typescript
client, which highlighted a couple of small bugs in the implementation,
which has also been fixed in this commit.

We also ensure that the generated client passes the
strictest Typescript type-checks and ESlint rules to ensure the
client doesn't cause any errors in peoples projects no matter how
strictly they've configured the project.