Geneated Clients: URI Encode Path Segments
This commit adds URI encoding to path segments when called
from the generated clients. This allows for user input
to be passed directly to the client and through the URL segments
without worrying about segments not aligning.

This work showed a bug where Encore's routing layer was using
the already decoded path, which meant the path:

```
/account/abc%2Fdeposit_usd%2F%2F200
```

Was being routed as:

```
/account/abc/deposit_usd/200
```

This was fixed by using `req.URL.EscapedPath()` in the runtime
and then adding `UnescapePath` to the code which reads the path
segments and passes them to the running application