Generated Clients: Error on missing response headers
This commit changes the generated clients, so when a header
is not sent by the server, but is expected by the client - rather
than providing a default value, the client will instead raise an
error.

This aligns with the behaviour when JSON payloads are not able
to be unmarshalled into the correct types and provides a more
fail-safe behaviour on the clients.