all: track known apps persistently
To support dependency injection Encore needs to generate code that is
written into the app's file tree so that editors can pick it up.

It would be frustrating if the code generation was only enabled after
running 'encore run', but the daemon had no way of knowing the location
of Encore apps on the local filesystem.

Fix this by introducing a local sqlite db that tracks known Encore apps
on the local filesystem. We can then use this to watch the filesystem for
changes to the known Encore apps and trigger code generation even without
having to run 'encore run'.