compiler/internal/codegen: Skip auth validation when empty auth params
This change adds tracking of decoded auth fields and returns an Unauthenticated error when there are no auth fields available.
For public endpoint this will skip the auth validation and proceed. For auth endpoints it'll return a "missing auth param" error.

If an auth param is available and any other error than Unauthenticated is returned, this request will fail.