Refactor resource service startup
This commit refactors the resource service startup, such
that if you add code for a SQL database or pubsub topic when
`encore run` is already running, the watch will pick up those
changes and start the required services on your system.

This involved:
- Moving the resource startup code into a new `ResourceServices` struct
- Moving the `OpTracker` into it's own package from `cli/daemon/run` so it could be passed around
- Making `OpTracker` reciever functions work will `nil` pointers (i.e. no-op) so the calling code doesn't need to check

While refactoring this code, I added a `asyncBuildJobs` which allows jobs which can be run in paralel to be run in paralel.

I also added an NSQ log adapter and limited the logs to warning's and above.