Add tool for generating Public API: encore.dev
This commit adds a tool which allows us to easily
generate the [encore.dev](https://github.com/encoredev/encore.dev)
public API from the runtime implementation.

This auto generation means we can be sure to easily keep the API in-step
with the actual implementation, without missing docs, comments or constants.