Update trace parsing
This commit updates the trace parsing to bring it inline with the
protobuf we use for traces in the long term storage engine and allows
the Parse function to be called from our engine directly.

This removes the need for us to maintain two copies of this proto and
parsing logic.