Fixed nil pointer on unbound resources.
Most requires are bound to package level varaibles, like this:
```go
var db = sqldb.Named("svc")
```

We track the identifier `db` as being bound to the database `svc`.
However, it's possible to make calls like this:

```go
var db = sqldb.Named("svc").Stdlib()
```

In this case `db` is not longer bound against the database `svc`, as
the named resource isn't directly bound.

The refactored resource parser, correctly identified the named database,
and it wasn't bound to a varaible in the package. However, older code
was written expecting an identified resource to always have an ident.

This is what caused the nil pointer panic. I've added checks for `nil`
identifiers, and skip the resource usage tracking if we know that resource
isn't bound against anything.