Refactor PubSub Subscription API design
This commit refactors the the PubSub subscription API to make
it a package level function, which aligns more with the concept
of resource creation.

This refactor required modifying the parser to introduced phased based
passes into the resource parser, such that Topics get parsed before
subscriptions as subscriptions need to reference the already parsed topics.
Without this passes approach, a subscription in `a.go` might be parsed before
the topic it references in `b.go`.

In a future PR more information will be validated on phase of topics and subscriptions
- such as forcing the `DeliveryGarauntees` option to be explicility set.

This commit also fixes the `publicapigen` to work for the PubSub package.