Refactor PubSub Subscription Handler API
This commit refactors PubSub subscription handlers to be
parsed into the subscription via the `pubsub.SusbcriptionConfig`
struct.

Parse time checks have been added to ensure that expect fields
are set within the `SubscriptionConfig` and `TopicConfig` structures
as compile time literals.

Finally we've changed the requirement for topic and subscription names
to align with the cronjob ID format (`kebab-case` and between 1 & 63
characters).