Compile time evaluation of PubSub Config
This commit adds the ability for Encore's parser
to evaulate the configuration struct for both
Topics and Subscriptions as compile time constants.

It means the compiler can then attach all the configuration
option data into the metadata for the application, which will
then allow the backend to provision the infrastructure as required.