Fix Local Dashboard Logging (#308)
This commit fixes encoredev/encore.dev#4 - which reported that the dashboard
log viewer does not render booleans correctly. This was due to the UI returning
a `true` or `false` value directly as a React.Node - which React does not render.

The fix was to render everything as JSON strings, which prevents thus bug, and
other falsy values from not rendering.

This commit also fixes an issue where if you're running multiple apps at the same time,
the home screen of the dash would not swap between the two apps when you use the app
switcher. The fix was to introduce a key for both the API caller and Trace List, such that
when the `AppID` changes, both those components get unmoutned and remoutned.