Add a how-to guide for using ent and atlas
This is based on the discussion here, and my own use of ent in my encore application: https://github.com/encoredev/encore/discussions/307.

With some tinkering, it's very easy to integrate a heavy ORM like ent with Encore. This guide shows one way of doing it, which could be expanded, but this seems to work really well on my end. We could also add a section for using only Atlas if this could be useful.

Any feedback would be very much appreciated!