cli: compiler: introduce user-facing codegen for DI
This adds the ability for Encore to generate code that is intended
to be written to Encore users' application directories. The motivating
use case is dependency injection, which involves generating and maintaining
a Client interface that mirrors the service's API.

To enable the functionality the experimental code must be enabled using
the new ENCORE_EXPERIMENT environment variable. Set ENCORE_EXPERIMENT=di
to test the new functionality during development.