runtime: add API description to metadata
In order for middleware to be able to generate
type-safe responses without having to call the
underlying handler, we need some way to access
what the type should be. Add an API description
struct to the metadata APIs that includes `reflect.Type`
values for the request/response types for such use cases.