Import all services into the main file
This commit ensures that all services get imported into
the generated `main.go` file for a built app.

This was to fix a bug where a service could be created using
a PubSub subscriber, but if a service only had subscribers and
had no API handlers, it was never imported by the main package,
thus never started up.