Add CUE file generator (#347)
* Add CUE file generator

This commit adds the ability for Encore to generate CUE files based on calls to `config.Load`
within their services. The generated CUE files can be used with the standard CUE tooling to
create and validate the config needed for the app.

We support both `cue` and `json` tags being used on config structs.

* Address Stefan's PR comments around CUE geneartion

I've also tieded up some of the formatting of the generated code.