Support Unit Tests in GoLand
This commit adds basic support for running tests from within
GoLand using the Encore GoLand Plugin (yet to be released).

GoLand runs tests in two phases; first it compiles the test, and
then it runs the compiled test file. It does this using the `-c`
argument to `go test`.

This causes problems for Encore as the current assumption is test
binaries will always execute through the Encore daemon, which isn't
the case here.

If we detect that `-c` has been passed to `encore test` then we need
to embed into the binary any enviromental varaibles that would normally
be passed from the daemon to the test (such as the secrets or runtime config).
This allows the Encore runtime to startup correctly for the tests when the execute.

A small change was also made to logging, such that when running in this way
we get human readable logs outputted, rather than the raw JSON payloads. Again
this was due to the fact the daemon is normally the entity receiving the logs
and it internally pretty prints them for the end users.