[ENC-796] Add support for Enviromental selectors
This commit adds support for enviromental selectors to the CUE
configuration system allowed in Encore applications.

We've taken the decision to embed the configuration in a definition
called `#Meta` which can easily be referenced, but shouldn't
interfer too much with the end users own CUE definitions.

We've also moved the applications own config into a `#Config`
definition which is closed. That `#Config` is then inlined, which
right now has no behaviour difference to how we where doing it.
However in an upcoming version of CUE, an inlined definition which
is considered closed will cause an error _if_ a field is added to
the config file which is not present in the closed definition.
i.e. this will detect typo's.