[ENC-796] Add support for Environmental selectors (#405)
This commit adds support for environmental selectors to the CUE
configuration system allowed in Encore applications.

We've decided to embed the configuration in a definition called `#Meta`
which can easily be referenced but shouldn't interfere too much with the
end users' own CUE definitions.

We've also moved the application's own config into a `#Config`
definition which is closed. That `#Config` is then inlined, which right
now has no behaviour difference from how we were doing it. However, in
an upcoming version of CUE, an inlined definition will be considered
closed and will cause an error _if_ a field is added to the config file
which is not present in the closed definition. i.e. this will detect
typos.

With this change, the follow configuration resulted in `encore run`
logging `932` for `FieldC`, however `encore test` on the exact same
code, resulted in `FieldC` logging `1337`.

```cue
FieldC: [
	if #Meta.Environment.Type  == "test"  { 1337 },
	if #Meta.Environment.Cloud == "local" { 932 },
	42
][0]
```

![Screenshot 2022-10-03 at 15 59
33](https://user-images.githubusercontent.com/236641/193625231-7829f645-af9d-4a76-9f7e-1205c92e03d7.png)

**Breaking Changes**

This commit removes the `EnvironmentType` `EnvLocal` and adds a new
`EnvTest` from the meta data API. This allows an Encore app to
understand if it's running within a unit test or not. I believe this
makes more sense as an app running under `encore run` will now have an
environment of `EnvDevelopment` and a `CloudProvider` of `CloudLocal`.
However running under `encore test`, it will be `EnvTest` &&
`CloudLocal`.