[ENC-800] Error if `config.Value` types are nested
This commit adds a new parser check to validate we don't have
nested `config.Value[config.Value[T]]` types, as that could lead
to unexpected behaviour.

Also fixes the test data for the `cuegen` as it hadn't
been updated.

Finnaly I've re-introduced teh `EnvLocal` constant, but marked it
as deprecated.