[ENC-799] Error on Unexported Config Fields
If a data type has unexported fields and is used by a call to
`config.Load[T]()` then we will now report an error to the user.
This error is to prevent undefined behaviour when a user expects
to be able to load an unexported field and is not able to.

This error will only occur on unexported fields used by config,
unexported fields used by any other Encore feature will be unaffected.