[ENC-834] Next Generation Error Messages
This commit massively changes how Encore renders error messages
to the end users in the CLI. It achieves this through a new package
`errinsrc` and two types in that package;

- `errinsrc.ErrInSrc` represents a single error found in the source code
- `errinsrc.List` represents a list of errors found in the source code

`errinsrc.ErrInSrc` will render the surrounding code of an error, and is
capaible of storing multiple source code locations witin a single error.
Each location can be marked as an Error, Warning or Help location and
can be anotationed with some text.

Stack traces _can_ be captured as well. By default stack traces will only
be captured for unhandled panics, but if encore is built with the `dev_build`
tag, then all errors will capture a stack. This allows for easier debugging
while working on Encore itself.

It then modifies the `errlist` package to use the new error types, while
maintaining the existing behaviour on the old methods. It achieves this
by upgrading all `scanner.Error` errors into a generic Go parser error.

All instances of a `ErrInSrc` are created in the `srcerrors` package,
 which contains a bunch of helper code to make it easy to create, but
 also keeps all our errors in one place, which allows easy mass updating
 (such as a possible future internationalation of the encore app).

Within `srcerrors` there are three generic errors which are kinda
catch-all types.:
- `UnhandledPanic`; for handling panics
- `GenericGoParserError`; for handling issues from the Go parser & scanner packages
- `GenericGoCompilerError`; for handling issues reported by the Go compiler

All other errors should have a custom type.

If this commit works as intended, end users should only ever see an
error rendered by `ErrInSrc`.