Allow `encore test` to support `-json`.
This commit allows encore to obey the `-json` arg
which instructs the tests to output everything in
raw JSON. Currently we where always pretty printing
the JSON output, which broke tooling intergration
(such as the GoLand plugin).

This works by refactoring our command streaming process
to take the concept of a `outputConverter` which operates
on a per line basis.

I've introduced two converters, the first one `convertJSONLogs()`
operates the same way as our original `streamCommandOutput` with
`convertJSON` set to true.

The second converter, `convertTestEventOutputOnly(base)` will decode
a Go `testEvent` and then call the `base` converter only on `Output`,
and then converts the updated `testEvent` back to JSON for tooling
to use and read.