[ENC-879] Fix chicken & egg issue with Cronjobs on Service Structs
This commit fixes an issue where using a service struct API as a
cronjob endpoint would result in a parse error if the `encore.gen.go`
files where missing from the file system.

This was due to the name resolution process needing not seeing the
function declartion, so not tracking `Foo` as a valid ident. Then
the cronjob parsing code would fail, as it failed to resolve `Foo`
to the API defined as `(*Service).Foo`.

This cronjob parser error then prevented `encore.gen.go` from being
generated, which would allow the parser to then parse successfully.

I've removed the `encore.gen.go` from the end to end app, and added
a cronjob to it to test the changes.