Add a JavaScript client to the code generation
For older or client who do not use TypeScript, having a JavaScript client generated with code similar to the TypeScript client could be fairly useful. I have tried to implement that JavaScript client while staying as true to the coding style of the TypeScript client as possible. I have tested it locally and implemented the tests to make sure it generates the clients as expected. Indentation was a bit annoying in parts, but I think this works pretty great for what it is.

Some alternative solutions:

- For creating the TypeScript client, we could also use the TypeScript compiler to compile the TypeScript client to JavaScript instead.
- For formatting, we could write teh code non-formatted and format it with a binary tool like [Rome](https://rome.tools/), which might work better than trying to ident it manually.

Closes #371