parser: fix method API name resolution
For Encore APIs defined as methods Encore generates
a user-facing `encore.gen.go` file that includes package-level
func wrappers for calling that API. That file is not intended
to be checked in to source control.

As a result we ended up with subtle bugs where the parsing
behavior depended on whether or not that file existed.
It existed during local development but not in CI/CD.

Fix this by ignoring that file during parsing to ensure consistent
behavior, and introduce a synthetic node in the name resolution
algorithm for method-based APIs to fix the underlying problem.