[ENC-888] Improve NG-Error Rendering on Client.
This commit improves the next generation errors rendering on the CLI.
It achieves this by introducing a new Command between the daemon and
the front end called display error. Whenever we want to display an
error list, this command is used to send the full `errlist.List` object
to the front end, which then renders it to `stderr`.

The reason we render the error list on the front end is:
1. We can calculate the size of the terminal window and then format
   the error to fix perfectly within the terminal
   (this is done by changing `errinsrc.ErrorWidth` before render).
2. If the command is run with `--verbose`, then only the client knows
   that, by rendering client side, `errlist` can obey the verbose command.

In the future:
1. Using bubbletea, we can expose a terminal UI for
   paging between multiple errors (when it shows `(and x more errors)`)
2. We can render errors nice the local dashboard within the browser.