[ENC-889] Use original files for compile errors (#436)
This commit fixes an issue where for compiler errors, we where using the
file which had been passed to the compiler. However due to `/*line
:blah` directives inserted by Encore during the rewrite process, the
rendered error lines did no align with the original source.

To fix it, we check if the file exists within the original app folder,
if it does we rewrite the filename to that file, if it doesn't (i.e.
it's an Encore generated file) we leave the filename as it was.