cli/daemon/sqldb: cache cluster status
We've observed ~200ms connection establishment
times when using Encore's postgres proxy.
Profiling showed it came from calling docker
to get the container's port. Introduce a cache
that stores the most recently observed value
and periodically refetch it, instead of checking
it every time a connection is established.

This reduces the connection time from 200ms
to 35ms on my machine.