codegen: unify path param names
This rewrites the paths we feed to `httprouter` in order to enable
different paths to use different parameter names in the same position.

This is not allowed by httprouter because it allows looking up parameters
by name, but we only use the positional indices so we don't need this
functionality. As a result we can just rewrite the paths to work within
the limits of httprouter and get the best of both worlds.