[ENC-804] Remove sorting of Parser errors (#485)
This commit removes the code which caused the parser to "sort" orders.
While this seems like it made sense, the issue is the order in which the
parser encounters errors is important as the first error can cause an
inconsistent state within the parser, which results in more errors being
detected and reported.

This was evident in the usecase where you had an API being referenced by
a pubsub subscriber. If you caused the API to have some parser failure;
such as an invalid signature, the PubSub subscriber would then report
that you can't reference the API in that way.

The problem was in some examples, the PubSub error would be sorted
first, resulting in the user thinking that PubSub can't access an API
directly - where in reality they just had an issue with the way one of
the API's was written.