compiler: skip generating test code if no tests exist
The go compiler uses the presence of `_test.go` files
as a signal to compile the test binary for a package.

Encore unconditionally generated such a file, which
meant that all packages would be compiled even if
they contained no actual tests. Fix this by only generating
such a file if the package contains other test files.