Disable JavaScript client generation for now
When I was investigating to make sure the JavaScript client was the same as the TYpeScript client, I noticed that the JS PR was merged before the JS client generation was properly disabled. According to the comment on my initial PR, I've created this follow up to make sure it's disabled if it's deemed necessary.