[ENC-1069] Support Private Network Access in CORS
As of Chrome 107, a new CORS header is present if the requested
resource is on a private network: `Access-Control-Allow-Private-Network: true`
The response _must_ include this header as well for the preflight to pass.

This commit adds support for this only on applications running locally
(i.e. run through `encore run`).

This commit also adds support for `encore.app` files to declare
a list of allowed headers which should be permitted on the application.
Previously this was configuration within the Encore platform but did not
affect local development.

By adding the following to `encore.app`, an app can opt into allowing the
header `X-Test-Header`:

```json
{
    "global_cors": {
        "allow_headers": [
            "X-Test-Header",
        ],
    },
}
```