[ENC-1072] Fix over watching of the filesystem
The new FSnotify library we introduced in v1.10.5 had a bug where it would
recursively recursive and watch the file system (i.e. we expontentaionally
watch the file system).

This resulted in two side effects:
 - Our attempt to ignore `node_modules` got bypassed, so we ended up
   watching everything inside node_modules
 - We spent a lot of time going over the same folders again and again

This has been fixed by moving the main logic into the `WalkDir` call within
the watch library.

Then we've modified the `walkDir` in the parser to also ignore the
same set of folders, so we walk the app as quickly as possible.

This commit also updates the `eerror` package and intergrates it better
with zerolog, so if an error with a stack trace  or an error
with meta data is logged, we will extract them and add the meta data as
new parameters to the log line, and add the stack trace under the log
line. (This change will also work for `srcerrors` stack traces too as
they both implement the same stack trace interface)