storage/sqldb: add sqldb.Error (#541)
This adds a `*sqldb.Error` type which exposes information about the
underlying database error. It also adds the new
`encore.dev/storage/sqldb/sqlerr` package with common error codes that
can be tested against in a database-agnostic way.

The error also contains the `DatabaseCode` field that provides a
non-portable way of getting the precise error from the underlying
database server.