Add GraphQL Operation Tracing
This commit adds the concept of a GraphQL operation event to our tracing implementation. This will allow the trace list view to show the name of the query, mutation or subscription to be shown rather than just the GraphQL endpoint name.

To achieve this, we attempt to parse a raw endpoints request body as a GraphQL query document (using the same parser that `gqlgen` uses). If this parse succeeds, we assume that the endpoint is processing GraphQL and add the trace data