publicapigen: rework panics
Unconditional panics cause static analysis tools to treat
the function as panicking (unsurprisingly). This in turn
means code following a call to an Encore SDK will get
marked as unreachable, which is a very confusing
and frustrating experience as a user.

Fix this by making the panic conditional, and adding
a naked return so the new code actually compiles.

Finally add a check that the generated code compiles.