[ENC-1288] Fix indeterministic build/encore run
This commit fixes a bug introduced in #505 when we made the parser run in parallel.

The issue ultimately was a race condition within the main goroutine, such that when the last worker finished parsing a package, the main goroutine could either read from the `pkgCh` or the `workerDone` channel.

If it picked the `workerDone` channel, then the last package would never be added to the list of known packages, which could result in random errors later in the encore parse/compile pipeline.

I've also updated the unit tests to ensure the test cache is cleared for each PR