Support resources define in test files
Support defining resources in test files even
if the test files themselves aren't part of a service.

Fix this by refactoring the logic to move away from
resource.Package (which is now removed) in favor of
resource binds, which keep track of the file it's declared in.
Tweak golden setup