Add new trace protocol protos
This adds the protobuf definitions and the parser
for the new streaming trace protocol.

The new trace protocol lays the ground work for gRPC support
by changing the protocol to stream events as they happen instead
in a stateless fashion. The previous protocol in contrast was
stateful and difficult to parse in a streaming fashion.